rt: Move test functions to rust_test_helpers.cpp
